% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_adaptive.R
\name{compute_metrics}
\alias{compute_metrics}
\title{Compute Performance Metrics}
\usage{
compute_metrics(X, B, S, A, w, lambda, delta)
}
\arguments{
\item{X}{Mixture matrix (N x D)}

\item{B}{Dictionary matrix (K x D)}

\item{S}{Data matrix (N x K)}

\item{A}{Incidence matrix (m x N)}

\item{w}{Edge weights (length m)}

\item{lambda}{Regularization parameter}

\item{delta}{Smoothing parameter}
}
\value{
List of metrics
}
\description{
Calculates various metrics to assess solution quality.
}
\keyword{internal}
