% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_graph.R
\name{project_to_simplex}
\alias{project_to_simplex}
\title{Project Vector to Simplex}
\usage{
project_to_simplex(v, target_sum = 1)
}
\arguments{
\item{v}{Vector to project}

\item{target_sum}{Target sum (default: 1)}
}
\value{
Projected vector
}
\description{
Projects a vector onto the probability simplex using efficient sorting.
}
\keyword{internal}
