% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialTV-package.R
\docType{package}
\name{spatialTV-package}
\alias{spatialTV-package}
\alias{spatialTV}
\title{spatialTV: Spatial Total Variation Regularization for Transcriptomics}
\description{
The spatialTV package provides tools for:
\itemize{
  \item Spatial cell type deconvolution with TV regularization
  \item Low-rank dictionary learning for cell type mixtures
  \item Adaptive edge weight learning via linear programming
  \item Efficient C++ implementation via Rcpp/RcppArmadillo
}
}
\details{
Implements spatially-aware cell type deconvolution for spatial transcriptomics
data using Total Variation (TV) regularization. The package uses the Fast
Iterative Shrinkage-Thresholding Algorithm (FISTA) for efficient optimization.
}
\section{Main Functions}{

\itemize{
  \item \code{\link{fit_spatial_mixture}}: Basic FISTA optimization with fixed weights
  \item \code{\link{fit_spatial_mixture_adaptive}}: Alternating optimization with learned weights
  \item \code{\link{build_spatial_graph}}: Construct Delaunay triangulation
  \item \code{\link{initialize_dictionary}}: Initialize dictionary matrix
  \item \code{\link{evaluate_objective}}: Evaluate objective function
}
}

\section{Mathematical Formulation}{

The optimization problem is:
\deqn{\min_{X,B} -\sum_{i=1}^N \log z_i + \lambda \sum_{e=1}^m w_e \sum_{d=1}^D \sqrt{(AX)_{ed}^2 + \delta}}
subject to simplex constraints on X and B, where:
\itemize{
  \item X: per-spot mixture proportions (N x D)
  \item B: cell type dictionary (K x D)
  \item A: spatial graph incidence matrix (m x N)
  \item w: edge weights
  \item \eqn{z_i = \sum_k S_{ik}(XB^T)_{ik}}: per-spot evidence
}
}

\references{
Beck, A. and Teboulle, M. (2009). A Fast Iterative Shrinkage-Thresholding
Algorithm for Linear Inverse Problems. SIAM Journal on Imaging Sciences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/liutom445/spatialTV}
  \item Report bugs at \url{https://github.com/liutom445/spatialTV/issues}
}

}
\author{
\strong{Maintainer}: Tom Liu \email{liutom@umich.edu}

Other contributors:
\itemize{
  \item University of Michigan [copyright holder]
}

}
\keyword{internal}
